+++ R/basic.R
@@ -6,8 +6,9 @@
 #' @param h Forecasting horizon
 #' @param level Confidence level for prediction intervals
 #' @param method forecasting method, either "mean", "median", or random walk ("rw")
-#' @param type_pi type of prediction interval currently, "gaussian", "bootstrap" or "blockbootstrap"
-#' @param block_length length of block for circular block bootstrap (\code{type_pi == 'blockbootstrap'})
+#' @param type_pi type of prediction interval currently, "gaussian", "bootstrap",
+#' "blockbootstrap" or "movingblockbootstrap"
+#' @param block_length length of block for (circular) "blockbootstrap" or "movingblockbootstrap"
 #' @param seed reproducibility seed for \code{type_pi == 'bootstrap'}
 #' @param B Number of bootstrap replications for \code{type_pi == 'bootstrap'}
 #'
@@ -56,7 +57,10 @@ basicf <- function(y,
                    h = 5,
                    level = 95,
                    method = c("mean", "median", "rw"),
-                   type_pi = c("gaussian", "bootstrap", "blockbootstrap"),
+                   type_pi = c("gaussian",
+                               "bootstrap",
+                               "blockbootstrap",
+                               "movingblockbootstrap"),
                    block_length = NULL,
                    seed = 1,
                    B = 100)
@@ -117,7 +121,7 @@ basicf <- function(y,
     return(structure(out, class = "mtsforecast"))
   }
 
-  if (type_pi %in% c("bootstrap", "blockbootstrap"))
+  if (type_pi %in% c("bootstrap", "blockbootstrap", "movingblockbootstrap"))
   {
     sims <- vector("list", length = B)
     for (i in 1:B)
@@ -125,16 +129,27 @@ basicf <- function(y,
       # sampling from the residuals
       set.seed(seed + i*100 + nchar(method))
 
-      idx <- switch(type_pi,
-                    bootstrap = sample.int(n = nrow(resids),
-                        size = h, replace = TRUE),
-                    blockbootstrap = mbb(
-                      r = resids,
-                      n = h,
-                      b = block_length,
-                      return_indices = TRUE,
-                      seed = seed + i*100 + nchar(method)
-                    )
+      idx <- switch(
+        type_pi,
+        bootstrap = sample.int(
+          n = nrow(resids),
+          size = h,
+          replace = TRUE
+        ),
+        blockbootstrap = mbb(
+          r = resids,
+          n = h,
+          b = block_length,
+          return_indices = TRUE,
+          seed = seed + i * 100 + nchar(method)
+        ),
+        movingblockbootstrap = mbb2(
+          r = resids,
+          n = h,
+          b = block_length,
+          return_indices = TRUE,
+          seed = seed + i * 100 + nchar(method)
+        )
       )
 
       sims[[i]] <- ts(as.matrix(fcast) + as.matrix(resids[idx, ]),